DHSCFT-767: Export line chart - move DomContainer to its own dir
