+++ frontend/fingertips-frontend/components/atoms/DomContainer/DomContainer.tsx
